requirements.txt
torch==2.6.0+cu124
--index-url https://download.pytorch.org/whl/cu124
transformers==4.51.3
peft==0.15.2
pillow==11.1.0
pandas==2.2.3
tqdm==4.67.1
bitsandbytes==0.45.5
